+++ src/views/FollowRequestList.vue
@@ -4,6 +4,22 @@
       <h1 class="content-header">
         {{ $t('follow_requests.follow_requests') }}
       </h1>
+      <div class="tab-bar">
+        <a
+          class="tab"
+          :class="{ active: tabName === 'incoming' }"
+          @click="switchTab('incoming')"
+        >
+          {{ $t('follow_requests.incoming') }}
+        </a>
+        <a
+          class="tab"
+          :class="{ active: tabName === 'outgoing' }"
+          @click="switchTab('outgoing')"
+        >
+          {{ $t('follow_requests.outgoing') }}
+        </a>
+      </div>
       <div v-if="!isLoading && profiles.length === 0" class="content-message">
         {{ $t('follow_requests.no_follow_requests_found') }}
       </div>
@@ -32,7 +48,7 @@
 import { onMounted, ref } from "vue"
 import { useI18n } from "vue-i18n"
 
-import { getFollowRequests } from "@/api/relationships"
+import { getFollowRequests, getOutgoingFollowRequests } from "@/api/relationships"
 import { Profile } from "@/api/users"
 import Loader from "@/components/Loader.vue"
 import ProfileListItem from "@/components/ProfileListItem.vue"
@@ -46,29 +62,45 @@ const { getActorLocation } = useActorHandle()
 const { ensureAuthToken } = useCurrentUser()
 const { setPageTitle } = useTitle()
 
+const tabName = ref<"incoming" | "outgoing">("incoming")
 const profiles = ref<Profile[]>([])
 const nextPageUrl = ref<string | null>(null)
 const isLoading = ref(false)
 
-onMounted(async () => {
-  setPageTitle(t("follow_requests.follow_requests"))
+async function loadPage(pageUrl?: string) {
   isLoading.value = true
-  const page = await getFollowRequests(ensureAuthToken())
-  profiles.value = page.profiles
+  const authToken = ensureAuthToken()
+  let page
+  switch (tabName.value) {
+    case "incoming":
+      page = await getFollowRequests(authToken, pageUrl)
+      break
+    case "outgoing":
+      page = await getOutgoingFollowRequests(authToken, pageUrl)
+      break
+  }
+  profiles.value.push(...page.profiles)
   nextPageUrl.value = page.nextPageUrl
   isLoading.value = false
+}
+
+async function switchTab(name: "incoming" | "outgoing") {
+  tabName.value = name
+  profiles.value.length = 0
+  nextPageUrl.value = null
+  await loadPage()
+}
+
+onMounted(async () => {
+  setPageTitle(t("follow_requests.follow_requests"))
+  await switchTab(tabName.value)
 })
 
 async function loadNextPage() {
   if (nextPageUrl.value === null) {
     return
   }
-  const page = await getFollowRequests(
-    ensureAuthToken(),
-    nextPageUrl.value,
-  )
-  profiles.value.push(...page.profiles)
-  nextPageUrl.value = page.nextPageUrl
+  await loadPage(nextPageUrl.value)
 }
 </script>
 
@@ -84,6 +116,12 @@ async function loadNextPage() {
 
 .content-message {
   @include content-message;
+
+  margin-bottom: $block-outer-padding;
+}
+
+.tab-bar {
+  @include tab-bar;
 
   margin-bottom: $block-outer-padding;
 }