+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Show outgoing follow requests on "Follow requests" page.
+
 ### Fixed
 
 - Don't show "Repost" in repost counter tooltip while author list is loading.